+++ src/ray/raylet/scheduling/fixed_point.cc
@@ -10,7 +10,7 @@ FixedPoint::FixedPoint(double d) {
 
 FixedPoint::FixedPoint(int i) { i_ = (i * RESOURCE_UNIT_SCALING); }
 
-FixedPoint::FixedPoint(unsigned long i) { i_ = (i * RESOURCE_UNIT_SCALING); }
+FixedPoint::FixedPoint(uint32_t i) { i_ = (i * RESOURCE_UNIT_SCALING); }
 
 FixedPoint::FixedPoint(int64_t i) : FixedPoint((double)i) {}
